third_party/WebKit/LayoutTests/editing/pasteboard/insert-font-weight.html

<!DOCTYPE html>
<html>
    <body>
        <div id="container" contenteditable="true"></div>
        <div id="content"><span style="font-weight: lighter;">hello</span><span style="font-weight: bolder;">world</span></div>
        <script src="../../resources/dump-as-markup.js"></script>
        <script>
            Markup.description('This tests inserting elements with font-weight set to lighter and bolder values');

            var content = document.getElementById('content');
            Markup.dump(content, 'Content to insert');

            document.querySelector('div[contenteditable]').focus();
            document.execCommand('InsertHTML', null, content.innerHTML);

            Markup.dump('container', 'After insertion');
        </script>
    </body>
</html>
