third_party/blink/web_tests/editing/pasteboard/insert-font-weight-expected.txt

This tests inserting elements with font-weight set to lighter and bolder values

Content to insert:
| <span>
|   style="font-weight: lighter;"
|   "hello"
| <span>
|   style="font-weight: bolder;"
|   "world"

After insertion:
| <span>
|   style="font-weight: lighter;"
|   "hello"
| <span>
|   style="font-weight: bolder;"
|   "world<#selection-caret>"
